+++ TP_LABS.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.28307.779
+VisualStudioVersion = 15.0.28307.902
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TP_LABS", "TP_LABS\TP_LABS.csproj", "{BB88961F-861A-447C-8C1E-AD6189ADE278}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TP_LABS", "TP_LABS.csproj", "{BB88961F-861A-447C-8C1E-AD6189ADE278}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,6 +20,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D1C9BB55-4BD4-4EAC-BF7E-4FB0F3D34A2F}
+		SolutionGuid = {CF51F7FB-1D56-48E0-931A-D0ABB4418371}
 	EndGlobalSection
 EndGlobal
